changes made
There is an extra whitespace before the column name of COLUMN_PRODUCT_SUPPLIER_NAME that should be removed to fix the crash when displaying the data.
In the <application> tag of your manifest file, you specified the attribute fullBackupContent but you didn't pass the back rules. please consider adding the backup rules or remove this attribute to avoid compilation errors.
There are two unused variables that should be removed or utilized.